Update OCAModelParameterStore.cc
In the method WriteToOCADBFile: (1) the loop over the laserball sin wave parameters was starting with iPar = 1 and it was skipping the first parameter of the lbDistPtr array. Also, (2) the last parameter that was being printed to the output file did not correspond to a parameter concerning the laserball light distribution and (3) it was being printed twice. Changing the loop to start with iPar = 0 fixed the first two problems and changing the if statements corrected the last problem.