Laserball Binned Distribution Index
In OCAModelParameterStore::WriteToOCADBFile, the index (iTheta*nCThetaBins + iPhi) in the array of parameters and errors was incorrect. As it was, for iTheta = 0, the last index from the iPhi loop would be 35. When iTheta incremented to 1, with iPhi = 0, the array index would be 12 ( = nCThetaBins ), meaning that previous elements of the array are being used as output again.